edit: styles modal and header
