--- src/components/layout/Header/LogoutModal.module.css
@@ -1,96 +1,0 @@
-.modalContent {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  position: absolute;
-  top: 50%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  background: var(--modal-background);
-  border-radius: 16px;
-  outline: none;
-  width: 533px;
-  height: 387px;
-  text-align: center;
-}
-
-.modalOverlay {
-  position: fixed;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-    backdrop-filter: blur(7px);
-  background: linear-gradient(180deg, #355359 0%, #3b5d63 100%);
-}
-
-
-.modal {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-}
-
-.title {
-  font-family: var(--font-family-base);
-font-weight: 700;
-font-size: 30px;
-text-align: center;
-color: var(--main-black);
-}
-
-.text {
-  font-family: var(--font-family-base);
-font-weight: 400;
-font-size: 18px;
-text-align: center;
-color: var(--main-black);
-margin-top: 38px;
-}
-
-.logoutBtn {
-border: none;
-  border-radius: 8px;
-padding: 11px 79px;
-width: 410px;
-height: 44px;
-background: linear-gradient(180deg, #355359 0%, #3b5d63 100%);
-margin-top: 24px;
-}
-
-.logoutBtnTxt{
-  font-family: var(--font-family-base);
-font-weight: 500;
-font-size: 18px;
-text-align: center;
-color: var(--main-white);
-}
-
-.cancelBtn {
-  background-color: transparent; 
-  border: 1px solid var(--main-black);
-  border-radius: 8px;
-  /* padding: 11px 68px; */
-  width: 410px;
-  height: 44px;
-  cursor: pointer;
-  margin-top: 21px;
-}
-
-.cancelBtnTxt{
-  font-family: var(--font-family);
-  font-weight: 500;
-  font-size: 18px;
-  text-align: center;
-  color: var(--main-black);
-}
-
-.cancelBtn:hover,
-.cancelBtn:focus {
- border: 1px solid var(--main-white);
-}
-
-.cancelBtnTxt:hover,
-.cancelBtnTxt:focus {
- color: var(--main-white);
-}
