+++ src/index.css
@@ -1,3 +1,4 @@
 @import 'styles/reset.css';
 @import 'styles/variables.css';
 @import 'styles/base.css';
+@import 'styles/container.css';
